requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==12.1.0        # For Image handling
pypdf==3.17.0        # For PDF manipulation